KroneckerTools

KroneckerTools computes chains of Kronecker products as described in Kamenik (2005).

The following computations are performed

c = (I<sub>p</sub> ⊗ A ⊗ I<sub>q</sub>)*b

c = (I<sub>p</sub> ⊗ A<sup>T</sup> ⊗ I<sub>q</sub>)*b

c = (A ⊗ A ⊗ ... ⊗ A)*b

d = (A<sup>T</sup> ⊗ A<sup>T</sup> ⊗ ... ⊗ A<sup>T</sup> ⊗ B)*c

C = A * (B ⊗ B ⊗ .... ⊗ B)

D = A * B * (C ⊗ C ⊗ .... ⊗ C)

D = A<sup>T</sup> * B * (C ⊗ C ⊗ .... ⊗ C)

E = AB(C ⊗ D ⊗ ... ⊗ D)

Installation

 julia> using Pkg
 julia> Pkg.add("KroneckerTools")

Algorithms

We exploit the following property of the Kronecker product: vec(A * B * C) = (C<sup>T</sup> ⊗ A) * vec(B), so as never to form the matrix corresponding to the Kronecker product and whenever possible use matrix product instead.

Let A, a m * n matrix, B, a matrix whose size depends on the context, and b = vec(B). It follows that

c = (I<sub>p</sub> ⊗ A) * b = vec(A * B), where B is a n * p matrix.

c = (A ⊗ I<sub>q</sub>) * b = vec(B * A<sup>T</sup>), where B is a q * m matrix

c = (I<sub>p</sub> ⊗ A ⊗ I<sub>q</sub>) * b is computed in p blocks

c<sub>i</sub> = vec(B<sub>i</sub> * A<sup>T</sup>), i = 1, ..., p where B<sub>i</sub> is the i<sup>th</sup> block of q * n elements of vector b

A chain of Kronecker products, (A<sub>1</sub> ⊗ A<sub>2</sub> ⊗ ... ⊗ A<sub>n</sub>) * b can be written as (A<sub>1</sub> ⊗ I<sub>p<sub>1</sub></sub>)* (I<sub>p<sub>2</sub></sub> ⊗ A<sub>2</sub> ⊗ I<sub>p<sub>2</sub></sub>)* ... * (I<sub>p<sub>n</sub></sub> ⊗ A<sub>n</sub>)b where b is a vector and p<sub>1></sub>, p<sub>2</sub>, ..., p<sub>n</sub>, q<sub>1</sub>, q<sub>2</sub>, ...,q<sub>n</sub> are such as making each group in brackets conformable.

References

O. Kamenik (2005), "Solving SDGE models: A new algorithm for the Sylvester equation", <i>Computational Economics 25</i>, 167–187.

Methods

KroneckerTools.a_mul_b_kron_c! — Method

a_mul_b_kron_c!(d::AbstractVecOrMat, a::AbstractVecOrMat, b::AbstractMatrix, c::AbstractMatrix, order::Int64)

Performs aB(c ⊗ c ⊗ ... ⊗ c). The solution is returned in matrix or vector d. order indicates the number of occurences of c. c must be a square matrix

We use vec(ab(c ⊗ c ⊗ ... ⊗ c)) = (c' ⊗ c' ⊗ ... ⊗ c' ⊗ a)vec(b)

KroneckerTools.a_mul_b_kron_c_d! — Method

a_mul_b_kron_c_d!(d::AbstractVecOrMat, a::AbstractVecOrMat, b::AbstractMatrix, c::AbstractMatrix, order::Int64)

Performs e = ab(c ⊗ d ⊗ ... ⊗ d). The solution is returned in matrix or vector e order indicates the number of occurences of d. c and d must be square matrices

We use vec(ab(c ⊗ d ⊗ ... ⊗ d)) = (c' ⊗ d' ⊗ ... ⊗ d' ⊗ a)vec(b)

KroneckerTools.a_mul_kron_b! — Method

a_mul_kron_b!(c::AbstractVector, a::AbstractVecOrMat, b::AbstractMatrix, order::Int64)

Performs a*(b ⊗ b ⊗ ... ⊗ b). The solution is returned in matrix c. order indicates the number of occurences of b

We use vec(a*(b ⊗ b ⊗ ... ⊗ b)) = (b' ⊗ b' ⊗ ... ⊗ b' ⊗ I)vec(a)

KroneckerTools.kron_at_kron_b_mul_c! — Method

function kron_at_kron_b_mul_c!(a::AbstractMatrix, order::Int64, b::AbstractMatrix, c::AbstractVector, work::AbstractVector)

updates c with (a^T ⊗ a^T ⊗ ... ⊗ a^T ⊗ b)c using c and work as work vectors

KroneckerTools.kron_at_kron_b_mul_c! — Method

function kron_at_kron_b_mul_c!(a::AbstractMatrix, order::Int64, b::AbstractMatrix, c::AbstractVector, offset_c::Int64, work::AbstractVector)

updates c at offset_c with (a^T ⊗ a^T ⊗ ... ⊗ a^T ⊗ b)c using c and work as work vectors

KroneckerTools.kron_at_kron_b_mul_c! — Method

function kron_at_kron_b_mul_c!(d::AbstractVector, a::AbstractMatrix, order::Int64, b::AbstractMatrix, c::AbstractVector, work1::AbstractVector, work2::AbstractVector)

computes d = (a^T ⊗ a^T ⊗ ... ⊗ a^T ⊗ b)c using work vectors work1 and work2

KroneckerTools.kron_at_kron_b_mul_c! — Method

function kron_at_kron_b_mul_c!(d::AbstractVector, offset_d::Int64, a::AbstractMatrix, order::Int64, b::AbstractMatrix, c::AbstractVector, offset_c::Int64, work1::AbstractVector, work2::AbstractVector, offset_w::Int64)

computes d = (a^T ⊗ a^T ⊗ ... ⊗ a^T ⊗ b)c using data from c at offsetc and work vectors work1 and work2 at offsetw

KroneckerTools.kron_mul_elem! — Method

kron_mul_elem!(p::Int64, q::Int64, m::Int64, a::AbstractMatrix, b::AbstractVector, c::AbstractVector)

Performs (Ip ⊗ a ⊗ Iq) b, where m,n = size(a). The result is stored in c.

KroneckerTools.kron_mul_elem_t! — Method

kron_mul_elem_t!(p::Int64, q::Int64, m::Int64, a::AbstractMatrix, b::AbstractVector, c::AbstractVector)

Performs (Ip ⊗ a' ⊗ Iq) b, where m,n = size(a). The result is stored in c.